Moved encryption from xiaomi to crypto
