+++ Cargo.toml
@@ -8,6 +8,7 @@ members = [
   "bin/isabel", # gateway between MQTT and temperature sensors
 
   "lib/alice", # Yandex.Alisa API types
+  "lib/crypto",
   "lib/bluetooth",
   "lib/inspinia", # Inspinia WS API types
   "lib/str_derive",
