+++ lib/crypto/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "crypto"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+aes = "0.8"
+cbc = "0.1"
+cipher = { version = "0.4", features = ["std", "block-padding"] }
+md-5 = "0.10"
+
+[dev-dependencies]
+hex-literal = "0.4"
