+++ lib/xiaomi/src/lib.rs
@@ -1,10 +1,8 @@
 mod device;
 mod discover;
 mod message;
-mod token;
 mod vacuum;
 
-pub use token::{parse_token, Token};
 pub use vacuum::{FanSpeed, Status, Vacuum};
 
 pub type ErasedError = Box<dyn std::error::Error + Send + Sync>;
